Split test_unshard.py into test_unshard/ directory
- test_parallel_info.py: TestNormalizeParallelInfo
- test_plan.py: TestComputeUnshardPlan (updated for Pydantic params)
- test_execute.py: TestExecuteUnshardPlan (updated for pick_world_ranks)
